+++ index.ts
@@ -295,6 +295,49 @@ export class Core {
 
         console.log(out);
     }
+
+    public static cmdOutLoadBalancer (loadBalancer : LoadBalancerOption) {
+        let out : string = "\n";
+        out += "<LoadBalancer>\n";
+        out += "type = " + loadBalancer.type;
+        out += "\n";
+        const head = " " + "number".padEnd(9) + " | " + "mode".padEnd(20) +  " | " + "proxy".padEnd(20);
+        let line = "";
+        for (let n = 0 ; n < head.length ; n++){
+            line += "-";
+        }
+        out += head + "\n";
+        out += line + "\n";
+
+        let number = 0;
+        for (let n = 0 ; n < loadBalancer.maps.length ; n++){
+            let map = loadBalancer.maps[n];
+            if (!map.clone) {
+                map.clone = 1;
+            }
+            else {
+                if (map.clone == "auto"){
+                    map.clone = os.cpus().length;
+                }
+            }
+            if (!map.proxy){
+                map.proxy = "";
+            }
+            for (let n2 = 0 ; n2 < map.clone ; n2++){
+                const td = " " + number.toString().padEnd(9) + " | " + map.mode.padEnd(20) + " | " + map.proxy.toString().padEnd(20);
+
+                let line = "";
+                for (let n = 0 ; n < td.length ; n++){
+                    line += "-";
+                }
+                out += td + "\n";
+                out += line + "\n";
+                number++;
+            }
+            number++;
+        }
+        console.log(out);
+    }
 }
 
 interface MineutServerInit {
@@ -372,10 +415,10 @@ export class MinuetServer {
                 console.log("# mkdir " + Core.initDir);
             }
 
-            //        if (!fs.existsSync(MinuetServerCore.initDir)){
+            if (!fs.existsSync(Core.initDir)){
                 fs.copyFileSync(__dirname + "/template/init.yaml", Core.initPath);
                 console.log("# make  init.yaml " + Core.initPath);    
-    //      }
+            }
         
             this.init = Core.getInit();
             console.log("# read init\n");
@@ -387,6 +430,8 @@ export class MinuetServer {
 
             const getLbServers = Core.getLbServers(this.init);
         
+            Core.cmdOutLoadBalancer(this.init.loadBalancer);
+            console.log("");
             Core.cmdOutSectors(sectors);
 
             new LoadBalancer({
@@ -395,16 +440,12 @@ export class MinuetServer {
                 workPath:  __dirname + "/server/worker",
                 servers: getLbServers,
             });
-            console.log("# Listen Start!");
+            console.log("\n# Listen Start!");
 
         }catch(error){
             console.log("[ERROR] : " + error.toString());
             console.log(error.stack);
         }
-    }
-
-    public static getSector() {
-
     }
 }
